Create Parent Classes for Objects
These folders hold the parent classes for a Farmer Object, a Seed Object, and a Season Object. Currently, they have limited functionality, but in future iterations we will add more methods and variables to these classes to create more robust classes that can be used as Objects for this Farmer Simulation.