Updated Farmer & Seed Class with MarketBuy Bug Fix
This commit has an updated farmer class in which the seedBag has been turned from an ArrayList to a Seed[] array. The Seed class also has an updated variable that hold the amount of harvestable crops of that particular seed. This commit also updates the bug that was found in the MarketBuy screen.