Updated Plant Growth With Time Advancement Functionality
With this update, the game finally supports plant growth functionality, which is coordinated by the time advancement functionality. For every day that is passed in the game, the plants in each plot will have their water reduced by 2 levels (to simulate the evaporation of water and make sure that the farmer sufficiently waters their plants EVERY day). If the crop is given too much or too little water, then it will die, else it will grow by one level for every day passed assuming that it isn't harvested or dead.