Farm Machinery Functionality
In this update, the game now limits the farmer to only harvest and water plants up to a certain maximum daily amount. If a farmer runs out of water or capacity to harvest, then they can go to the market and buy irrigation systems or a tractor to help them increase this maximum daily amount limit and improve their crop production.